Extract role edit logic into mixin
